Update yk-csr-gen options and add new features
This commit updates the options for the yk-csr-gen tool and adds new features for generating a Certificate Signing Request (CSR) using a private key from a YubiKey device. The `--new-pk` flag now accepts `False` as the default value. The `--slot` flag allows the user to specify the slot number on the YubiKey device, with the default value set to `9D`. The `--out` flag accepts an optional file path where the generated CSR will be written. Additionally, the `--out-pub` flag allows the user to specify a file path where the public key extracted from the CSR will be written. The `--text` flag now accepts `False` as the default value, and the `--text-pub` flag is introduced to print the public key to the console. The `--hash` flag allows the user to specify the hash algorithm, with the default value set to `SHA256`. Various other flags such as `--cn`, `--c`, `--dc`, `--e`, `--city`, `--organizational-unit`, `--o`, `--province`, `--s-dns`, `--s-e`, `--s-ip`, `--s-uri`, `--s-upn`, `--oid`, `--version`, `-?, -h, --help` have been updated or added to provide more flexibility and functionality to the tool.

Signed-off-by: Bryan Gonzalez <bryan@battleface.com>