Add release build workflow for archiving and hashing files
The release.yml file has been modified to add a new workflow named "Release Build" for archiving and hashing files. This workflow will trigger on release events of type "published" or "edited". In the workflow, a step has been added to create a ZIP archive of the target directory and generate a SHA256 checksum of the archive. Finally, the ZIP archive and checksum file are uploaded.

Signed-off-by: Bryan Gonzalez <bryan@battleface.com>